IMPALA-11738: Hide symbols from compression libraries for libfesupport.so
Recently, dataload has been failing on some configurations
with an error from Hive when initializing zlib native code
in ZlibDecompressor.init(). This error goes away when
libfesupport.so is removed from JAVA_LIBRARY_PATH in
testdata/bin/run-hive-server.sh, so something about
libfesupport.so is interfering with the functioning of
zlib.

libfesupport.so includes several compression libraries
(libbz2, liblz4, libsnappy, libz, libzstd). To avoid these
types of conflicts, this hides the symbols for the
compression libraries in libfesupport.so. That prevents
libhadoop from using the symbols in libfesupport.so.

Testing:
 - Ran an ASAN dataload on Centos 7 (which previously had
   been consistenly failing)
 - Ran precommit job

Change-Id: I55bda6899044ff2ad98134f5954df83f3e10a5cc
Reviewed-on: http://gerrit.cloudera.org:8080/19264
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>