IMPALA-11738: Add toolchain libstdc++ to LD_LIBRARY_PATH for HiveServer2
libfesupport.so is built against the toolchain libstdc++,
which may be newer than the OS libstdc++. Since HiveServer2
has libfesupport.so on its java.library.path, we should also
put the toolchain libstdc++ in LD_LIBRARY_PATH so that
libfesupport.so can resolve all its symbols properly.

This fixes an issue with a Hive crash in Ubuntu 18 dataload.

Testing:
 - Ran dataload on Ubuntu 18

Change-Id: Ib35e778850618b35073d102b89a4de02071b4a19
Reviewed-on: http://gerrit.cloudera.org:8080/19276
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>