IMPALA-11803: Fix hitting DCHECK when running union on empty table with MT_DOP>1
The DCHECK was caused as the value of useMtScanNode_
was not getting set for empty tables. This was because of
a condition for empty tables that returned from the
computeNodeResourceProfile() function before setting up
useMtScanNode_. It got rectified by placing the
assignment of useMtScanNode_ before that condition.

Testing: The fix is verified with an end-to-end test in test_mt_dop.py.

Change-Id: Idbae5e1a78211327a214b2d936743bda767ae3c4
Reviewed-on: http://gerrit.cloudera.org:8080/19474
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>