IMPALA-10971: Fix data cache metrics for instant evictions
When inserting an entry into the data cache, the
Insert() can fail and instantly evict the new entry.
This is currently only true for the LIRS cache
eviction policy. When this happens, the cache metrics
are not maintained properly. The instant eviction
results in an unmatched decrement to the total bytes
and num entries counters.

This moves the increments to happen prior to the
Insert() call, which fixes the instant eviction
case. The behavior does not change noticeably for
the successful Insert() case.

Testing:
 - Added custom cluster test to verify this case

Change-Id: I5db3eaca2f4459844e3270846d905f76265b6b3e
Reviewed-on: http://gerrit.cloudera.org:8080/19930
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>