IMPALA-11989: ImpalaRuntimeException if Kudu and Impala use different HMS
Added a startup flag 'enable_kudu_impala_hms_check' which is used to
flag an error if Kudu and Impala are using different HMS instance(s).
The default behavior in Impala is not changed and the default value of
the new flag is true.

The check itself could be improved in future to ensure that the backend
store used by HMS is same. For now allowing a config option to disable
the check where it makes sense seems like a good short term solution.

Change-Id: I2b06935b262e4a314ad631167c1cd33319e4b10f
Reviewed-on: http://gerrit.cloudera.org:8080/19605
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>