IMPALA-11818: Fix flaky TestDdlStatements.test_describe_materialized_view
This test is executed on multiple ImpalaDs serially. This caused some
race condition between the CREATE/DROP database statements, the test
could fail with "Database does not exist". This commit adds SYNC_DDL to
the test, so the CREATE/DROP db won’t return till the metadata is
propagated to each ImpalaDs.

Testing:
 - Could not trigger the failure locally, but ran the test

Change-Id: I399cb70d4819f27ecd752bf709a77e2e31c5ab54
Reviewed-on: http://gerrit.cloudera.org:8080/19395
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>