IMPALA-11825: Doc update 4.2.0 breaks web search results
We recently updated the upstream Impala docs in
https://gerrit.cloudera.org/#/c/19367/. The doc build system introduced
a new directory with CSS formatted html but also renamed the old 'html'
directory to 'plain-html'. The problem is that changing the directory
name from 'html' to 'plain-html' breaks web search results as they still
have the old links.

This change adds a symlink from 'html' to 'asf-site-html' to fix it.

Change-Id: Iab1d5fe5e3be85f26aeeb24be9e0018935d2e424
Reviewed-on: http://gerrit.cloudera.org:8080/19402
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>