IMPALA-11826: Avoid calling planFiles() on Iceberg V2 tables when there are no predicates
Similar to IMPALA-11591 but this Jira extends it to V2 tables. With
this patch we group data files into two categories in
IcebergContentFileStore:
 * data files without deletes
 * data files with deletes

With this information we can avoid calling planFiles() when planning
the scans of Iceberg tables. We can just set the lists of the file
descriptors based on IcebergContentFileStore then invoke the regular
planning methods.

iceberg-v2-tables.test had to be updated a bit because now we are
calculating the lengths of the file paths based on Impala's file
descriptor objects + table location, and not based on data file
information in Iceberg metadata (which has the file system prefix
stripped)

Testing:
  * executed existing tests
  * Updated plan tests

Change-Id: Ia46bd2dce248a9e096fc1c0bd914fc3fa4686fb0
Reviewed-on: http://gerrit.cloudera.org:8080/19419
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>