IMPALA-11829 - Fix bug in cardinality estimates related to TABLE_NUM_ROWS hint
IMPALA-7942 added support for a TABLE_NUM_ROWS query hint which can be used
to specify a table cardinality for cases where stats are missing or invalid.
In the case where stats were missing or invalid and no TABLE_NUM_ROWS hint was specified
by the user, HdfsScanNode.getStatsNumRows was incorrectly returning a default value of -1
instead of returning a rough estimate of cardinality as it had prior to the IMPALA-7942 change.
This change fixes the return value of getStatsNumRows so it only uses the TABLE_NUM_ROWS
value when the users has actually specified the query hint.

Change-Id: Ia27745fd93abd5dec99bf82f16899bd15a2b88ae
Reviewed-on: http://gerrit.cloudera.org:8080/19421
Reviewed-by: Qifan Chen <qfchen@hotmail.com>
Reviewed-by: wangsheng <skyyws@163.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>